+++ resources/js/components/About.jsx
