+++ resources/js/components/GlobalNav.jsx
